Review Changes
* Changed all CSP Nonce related code a separeate file
* Removed the shallow copy *r = *r.WithContext()
* Updated tests with the new changes
* Added withCSPNonce functions
* Renamed CSP related fucntions and variables to explicitly start with CSP
